split load_corpus and save_corpus modules
