+++ backend/addcorpus/tests/test_save_corpus.py
@@ -1,6 +1,6 @@
 from addcorpus.tests.mock_csv_corpus import MockCSVCorpus
 from addcorpus.models import Corpus
-from addcorpus.load_corpus import _save_field_in_database
+from addcorpus.save_corpus import _save_field_in_database
 
 def test_save_field_definition(db, mock_corpus):
     corpus = Corpus.objects.get(name=mock_corpus)
