+++ backend/conftest.py
@@ -2,7 +2,8 @@ import pytest
 from allauth.account.models import EmailAddress
 from time import sleep
 from ianalyzer.elasticsearch import elasticsearch
-from addcorpus.load_corpus import load_all_corpora, load_corpus
+from addcorpus.load_corpus import load_corpus
+from addcorpus.save_corpus import load_and_save_all_corpora
 from es import es_index as index
 
 # user credentials and logged-in api clients
@@ -82,7 +83,7 @@ def es_client():
 def django_db_setup(django_db_setup, django_db_blocker):
     #add mock corpora to the database at the start of each test
     with django_db_blocker.unblock():
-        load_all_corpora()
+        load_and_save_all_corpora()
 
 def index_test_corpus(es_client, corpus_name):
     corpus = load_corpus(corpus_name)
