+++ backend/media/conftest.py
@@ -1,5 +1,4 @@
 import pytest
-from addcorpus.load_corpus import load_all_corpora
 
 @pytest.fixture()
 def mock_corpus():
